Update index.html
A simple To-Do app. This application describes and allows the user to enter their daily to do things . the addTask() allow the user to add their task and the remove prompt allow the user to remove any task after they complete it. The javascript and Css files are accessed to the main html file . And enhanced the css for better user experience